move AccountListSidebar out of screens folder
