+++ src/renderer/ScreenController.tsx
@@ -18,7 +18,7 @@ import About from './components/dialogs/About'
 import { KeybindingsContextProvider } from './contexts/KeybindingsContext'
 import { DialogContext } from './contexts/DialogContext'
 import WebxdcSaveToChatDialog from './components/dialogs/WebxdcSendToChat'
-import AccountListSidebar from './components/screens/AccountListSidebar/AccountListSidebar'
+import AccountListSidebar from './components/AccountListSidebar/AccountListSidebar'
 import SettingsStoreInstance from './stores/settings'
 import { NoAccountSelectedScreen } from './components/screens/NoAccountSelectedScreen/NoAccountSelectedScreen'
 import AccountDeletionScreen from './components/screens/AccountDeletionScreen/AccountDeletionScreen'
