+++ src/renderer/components/AccountListSidebar/AccountItem.tsx
@@ -2,12 +2,12 @@ import React, { useEffect, useState } from 'react'
 import classNames from 'classnames'
 import debounce from 'debounce'
 
-import { BackendRemote, onDCEvent } from '../../../backend-com'
-import { avatarInitial } from '../../Avatar'
-import { getLogger } from '../../../../shared/logger'
-import useTranslationFunction from '../../../hooks/useTranslationFunction'
-import { useContextMenuWithActiveState } from '../../ContextMenu'
-import { ActionEmitter, KeybindAction } from '../../../keybindings'
+import { BackendRemote, onDCEvent } from '../../backend-com'
+import { avatarInitial } from '../Avatar'
+import { getLogger } from '../../../shared/logger'
+import useTranslationFunction from '../../hooks/useTranslationFunction'
+import { useContextMenuWithActiveState } from '../ContextMenu'
+import { ActionEmitter, KeybindAction } from '../../keybindings'
 
 import styles from './styles.module.scss'
 import type { T } from '@deltachat/jsonrpc-client'
