+++ src/renderer/components/AccountListSidebar/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../scss/variables';
+@import '../../../../scss/variables';
 
 .avatar {
   --local-avatar-size: 48px;
